Sidebar recipes live in the registry (carried on the mark definition)
Moves the 5 sidebar-card recipes into a shared src/lib/sidebar/recipe.ts (pure
data + types, no client/worker imports) and attaches each to its mark definition
in code-marks.ts. So a recipe-driven card's STRUCTURE now lives with its mark and
is exposed via /api/marks — the payoff of the generic-sidebar migration.

One source of truth: the client CARD_DEFS imports the same recipe objects (via
ArgumentSidebar's re-export), and the worker references them on MarkDefinition.
No duplication, no async fetch, behavior byte-identical (Codex confirmed the 5
recipe literals are unchanged from the old client consts).

The shared recipe type loosens two client-only narrowings to plain strings —
`kind` (accent() casts back to SidebarKind with a fallback) and explainer
`labelKey` (CatalogKey -> string; SectionCard's label widened; a new test asserts
every labelKey still resolves via t(), recovering the safety). MarkDefinition +
the client /api/marks type gain optional `recipe?`. recipeHash/cache keys are
untouched (they hash extractor[+render], not the whole def).

Codex-reviewed: no blocking findings; byte-equivalent recipes, clean lib (no
Solid pulled into the worker), cache paths unaffected.
